Notes READ.ME
This weeks project was a giphy website. I made mine a "Animal" specific website. On the website you have buttons with the name of the animal and when clicked 10 different giphs are generated related to the specific animal you chose. there also is a search bar where you can search for your own animals, type in animal and you will generate 10 animal giphs for that animal. Enjoy!!

I didn't have enough time to set the animal giphs to still animate and i would have made the website prettier if I had more time.